refactor: revert to imterface system
