+++ wallet-gateway/remote/src/ledger/wallet-allocation/signing-providers/dfns-wallet-allocator.ts
@@ -6,26 +6,40 @@ import { Store, UpdateWallet, Wallet } from '@canton-network/core-wallet-store'
 import {
     SigningDriverInterface,
     SigningProvider,
+    Error as SigningError,
 } from '@canton-network/core-signing-lib'
 import { Logger } from 'pino'
 import { PartyAllocationService } from '../../party-allocation-service.js'
 import { PartyHint, Primary } from '../../../user-api/rpc-gen/typings.js'
 import { WALLET_DISABLED_REASON } from '@canton-network/core-types'
-import { WalletAllocator } from './base.js'
+import { WalletAllocator } from '../wallet-allocation-service.js'
 
 /**
  * Dfns sign-only allocator. The gateway runs the topology flow against its
  * configured validator; Dfns is asked only to sign the topology hash, mirroring
  * the bring-your-own-validator pattern used for other external providers.
  */
-export class DfnsWalletAllocator extends WalletAllocator {
+export class DfnsWalletAllocator implements WalletAllocator {
     constructor(
         private store: Store,
         private logger: Logger,
         private partyAllocator: PartyAllocationService,
         protected signingDriver: SigningDriverInterface
-    ) {
-        super(signingDriver)
+    ) {}
+
+    private handleSigningError<T extends object>(result: SigningError | T): T {
+        if ('error' in result) {
+            throw new Error(
+                `Error from signing driver: ${result.error_description}`
+            )
+        }
+        return result
+    }
+
+    async getKeys(userId: UserId) {
+        if (!this.signingDriver) return null
+        const driver = this.signingDriver.controller(userId)
+        return await driver.getKeys().then(this.handleSigningError)
     }
 
     async createWallet(
